added logger to BicepNetTokenCredentialFactory, moved auth logic from BicepWrapper
